Idk what this is doing anymore, just seperated a bunch of classes from game.py
